Crate srml_im_online

I'm online Module

If the local node is a validator (i.e. contains an authority key), this module gossips a heartbeat transaction with each new session. The heartbeat functions as a simple mechanism to signal that the node is online in the current era.

Received heartbeats are tracked for one era and reset with each new era. The module exposes two public functions to query if a heartbeat has been received in the current era or session.

The heartbeat is a signed transaction, which was signed using the session key and includes the recent best block number of the local validators chain as well as the NetworkState. It is submitted as an Unsigned Transaction via off-chain workers.

im_online::Trait

Call

Module

Interface

Public Functions

is_online_in_current_session - True if the validator sent a heartbeat in the current session.

Usage

use srml_support::{decl_module, dispatch::Result};
use system::ensure_signed;
use srml_im_online::{self as im_online};

pub trait Trait: im_online::Trait {}

decl_module! {
	pub struct Module<T: Trait> for enum Call where origin: T::Origin {
		pub fn is_online(origin, authority_index: u32) -> Result {
			let _sender = ensure_signed(origin)?;
			let _is_online = <im_online::Module<T>>::is_online_in_current_session(authority_index);
			Ok(())
		}
	}
}

Dependencies

This module depends on the Session module.

Modules

ed25519

sr25519

Structs

GenesisConfig

Heartbeat
Heartbeat which is sent/received.

Module

UnresponsivenessOffence
An offence that is filed if a validator didn't send a heartbeat message.

Enums

Call

RawEvent
Events for this module.

Traits

Trait

Type Definitions

AuthIndex

Event
RawEvent specialized for the configuration Trait